+++ pom.xml
@@ -115,16 +115,10 @@
             <artifactId>okhttp</artifactId>
             <version>4.10.0</version>
         </dependency>
-
         <!-- Consumer  -->
         <dependency>
             <groupId>no.nav.tjenestespesifikasjoner</groupId>
             <artifactId>person-v3-tjenestespesifikasjon</artifactId>
-            <version>${tjenestespesifikasjoner.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>no.nav.tjenestespesifikasjoner</groupId>
-            <artifactId>egenansatt-v1-tjenestespesifikasjon</artifactId>
             <version>${tjenestespesifikasjoner.version}</version>
         </dependency>
 
@@ -206,7 +200,7 @@
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-contract-stub-runner</artifactId>
-            <version>3.1.2</version>
+            <version>3.1.3</version>
             <scope>test</scope>
         </dependency>
         <dependency>
