--- src/main/java/no/nav/veilarbperson/client/egenansatt/EgenAnsattClient.java
@@ -1,10 +1,0 @@
-package no.nav.veilarbperson.client.egenansatt;
-
-import no.nav.common.health.HealthCheck;
-import no.nav.common.types.identer.Fnr;
-
-public interface EgenAnsattClient extends HealthCheck {
-
-    boolean erEgenAnsatt(Fnr fnr);
-
-}
