+++ src/main/java/no/nav/veilarbperson/client/pdl/domain/Familiemedlem.java
@@ -2,8 +2,6 @@ package no.nav.veilarbperson.client.pdl.domain;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-import no.nav.common.types.identer.Fnr;
-import no.nav.veilarbperson.client.person.domain.RelasjonsBosted;
 
 import java.time.LocalDate;
