+++ src/main/java/no/nav/veilarbperson/client/pdl/domain/Sivilstand.java
@@ -2,7 +2,6 @@ package no.nav.veilarbperson.client.pdl.domain;
 
 import lombok.Data;
 import lombok.experimental.Accessors;
-import no.nav.veilarbperson.client.person.domain.RelasjonsBosted;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
