+++ src/main/java/no/nav/veilarbperson/client/person/PersonClientImpl.java
@@ -9,19 +9,12 @@ import no.nav.tjeneste.virksomhet.person.v3.binding.PersonV3;
 import no.nav.tjeneste.virksomhet.person.v3.informasjon.Informasjonsbehov;
 import no.nav.tjeneste.virksomhet.person.v3.informasjon.NorskIdent;
 import no.nav.tjeneste.virksomhet.person.v3.informasjon.PersonIdent;
-import no.nav.tjeneste.virksomhet.person.v3.informasjon.Personidenter;
 import no.nav.tjeneste.virksomhet.person.v3.meldinger.HentPersonRequest;
 import no.nav.tjeneste.virksomhet.person.v3.meldinger.HentPersonResponse;
-import no.nav.tjeneste.virksomhet.person.v3.meldinger.HentSikkerhetstiltakRequest;
-import no.nav.tjeneste.virksomhet.person.v3.meldinger.HentSikkerhetstiltakResponse;
-import no.nav.veilarbperson.client.person.domain.TpsPerson;
 import no.nav.veilarbperson.config.CacheConfig;
 import no.nav.veilarbperson.utils.MapExceptionUtil;
-import no.nav.veilarbperson.utils.PersonDataMapper;
 import org.apache.cxf.interceptor.LoggingOutInterceptor;
 import org.springframework.cache.annotation.Cacheable;
-
-import java.util.Optional;
 
 @Slf4j
 public class PersonClientImpl implements PersonClient {
@@ -36,10 +29,6 @@ public class PersonClientImpl implements PersonClient {
                 .build();
     }
 
-    public PersonClientImpl(PersonV3 personV3) {
-        this.personV3 = personV3;
-    }
-
     @Cacheable(CacheConfig.TPS_PERSON_CACHE_NAME)
     @Override
     public TpsPerson hentPerson(Fnr ident) {
@@ -48,20 +37,6 @@ public class PersonClientImpl implements PersonClient {
             return PersonDataMapper.tilTpsPerson(response.getPerson());
         } catch (Exception e) {
             log.error("Henting av person feilet", e);
-            throw MapExceptionUtil.map(e);
-        }
-    }
-
-    @Cacheable(CacheConfig.SIKKERHETSTILTAK_CACHE_NAME)
-    @Override
-    public String hentSikkerhetstiltak(Fnr fnr) {
-        try {
-            HentSikkerhetstiltakResponse wsSikkerhetstiltak = personV3.hentSikkerhetstiltak(lagHentSikkerhetstiltakRequest(fnr));
-            return Optional.ofNullable(wsSikkerhetstiltak.getSikkerhetstiltak())
-                    .map(no.nav.tjeneste.virksomhet.person.v3.informasjon.Sikkerhetstiltak::getSikkerhetstiltaksbeskrivelse)
-                    .orElse(null);
-        } catch (Exception e) {
-            log.error("Henting av sikkerhetstiltak feilet", e);
             throw MapExceptionUtil.map(e);
         }
     }
@@ -79,20 +54,7 @@ public class PersonClientImpl implements PersonClient {
     private HentPersonRequest lagHentPersonRequest(Fnr fnr) {
         return new HentPersonRequest()
                 .withAktoer(new PersonIdent().withIdent(new NorskIdent().withIdent(fnr.get())))
-                .withInformasjonsbehov(
-                        Informasjonsbehov.ADRESSE, Informasjonsbehov.BANKKONTO,
-                        Informasjonsbehov.FAMILIERELASJONER, Informasjonsbehov.KOMMUNIKASJON
-                );
-    }
-
-    private HentSikkerhetstiltakRequest lagHentSikkerhetstiltakRequest(Fnr fnr) {
-        PersonIdent personIdent = new PersonIdent();
-        personIdent.setIdent(
-                new NorskIdent()
-                        .withIdent(fnr.get())
-                        .withType(new Personidenter()
-                                .withValue("fnr")));
-        return new HentSikkerhetstiltakRequest().withAktoer(personIdent);
+                .withInformasjonsbehov(Informasjonsbehov.BANKKONTO);
     }
 
 }
