+++ src/main/java/no/nav/veilarbperson/client/person/TpsPerson.java
@@ -1,0 +1,12 @@
+package no.nav.veilarbperson.client.person;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+@Data
+@Accessors(chain = true)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class TpsPerson  {
+    String kontonummer;
+}
