--- src/main/java/no/nav/veilarbperson/client/person/domain/Diskresjonskoder.java
@@ -1,13 +1,0 @@
-package no.nav.veilarbperson.client.person.domain;
-
-public enum Diskresjonskoder {
-    STRENGT_FORTROLIG_ADRESSE("SPSF", "6"), FORTROLIG_ADRESSE("SPFO", "7");
-
-    public final String kodeverkVerdi;
-    public final String tallVerdi;
-
-    Diskresjonskoder(String kodeverkVerdi, String tallVerdi) {
-        this.kodeverkVerdi = kodeverkVerdi;
-        this.tallVerdi= tallVerdi;
-    }
-}
