--- src/main/java/no/nav/veilarbperson/client/person/domain/PostAdresse.java
@@ -1,18 +1,0 @@
-package no.nav.veilarbperson.client.person.domain;
-
-public class PostAdresse {
-    private UstrukturertAdresse ustrukturertAdresse;
-
-    public UstrukturertAdresse getUstrukturertAdresse() {
-        return ustrukturertAdresse;
-    }
-
-    public PostAdresse withUstrukturertAdresse(UstrukturertAdresse ustrukturertAdresse) {
-        this.ustrukturertAdresse = ustrukturertAdresse;
-        return this;
-    }
-
-    public PostAdresse copy() {
-        return new PostAdresse().withUstrukturertAdresse(ustrukturertAdresse.copy());
-    }
-}
