--- src/main/java/no/nav/veilarbperson/client/veilarbportefolje/Personinfo.java
@@ -1,7 +1,0 @@
-package no.nav.veilarbperson.client.veilarbportefolje;
-
-// Domeneobject for uthenting av data fra veilarbportefolje
-public class Personinfo {
-    public String sikkerhetstiltak;
-    public boolean egenAnsatt;
-}
