--- src/main/java/no/nav/veilarbperson/client/veilarbportefolje/VeilarbportefoljeClient.java
@@ -1,10 +1,0 @@
-package no.nav.veilarbperson.client.veilarbportefolje;
-
-import no.nav.common.health.HealthCheck;
-import no.nav.common.types.identer.Fnr;
-
-public interface VeilarbportefoljeClient extends HealthCheck {
-
-    Personinfo hentPersonInfo(Fnr fodselsnummer);
-
-}
