--- src/main/java/no/nav/veilarbperson/client/veilarbportefolje/VeilarbportefoljeClientImpl.java
@@ -1,56 +1,0 @@
-package no.nav.veilarbperson.client.veilarbportefolje;
-
-import lombok.SneakyThrows;
-import no.nav.common.health.HealthCheckResult;
-import no.nav.common.health.HealthCheckUtils;
-import no.nav.common.rest.client.RestClient;
-import no.nav.common.rest.client.RestUtils;
-import no.nav.common.types.identer.Fnr;
-import no.nav.veilarbperson.config.CacheConfig;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
-import org.springframework.cache.annotation.Cacheable;
-
-import java.util.function.Supplier;
-
-import static no.nav.common.utils.UrlUtils.joinPaths;
-import static org.springframework.http.HttpHeaders.ACCEPT;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-
-public class VeilarbportefoljeClientImpl implements VeilarbportefoljeClient {
-
-    private final String veilarbportefoljeUrl;
-
-    private final OkHttpClient client;
-
-    private final Supplier<String> userTokenProvider;
-
-    public VeilarbportefoljeClientImpl(String veilarbportefoljeUrl, Supplier<String> userTokenProvider) {
-        this.veilarbportefoljeUrl = veilarbportefoljeUrl;
-        this.userTokenProvider = userTokenProvider;
-        this.client = RestClient.baseClient();
-    }
-
-    @Cacheable(CacheConfig.VEILARBPORTEFOLJE_PERSONINFO_CACHE_NAME)
-    @SneakyThrows
-    @Override
-    public Personinfo hentPersonInfo(Fnr fodselsnummer) {
-        Request request = new Request.Builder()
-                .url(joinPaths(veilarbportefoljeUrl, "/api/personinfo/", fodselsnummer.get()))
-                .header(ACCEPT, APPLICATION_JSON_VALUE)
-                .header(AUTHORIZATION, "Bearer " + userTokenProvider.get())
-                .build();
-
-        try (Response response = client.newCall(request).execute()) {
-            RestUtils.throwIfNotSuccessful(response);
-            return RestUtils.parseJsonResponseOrThrow(response, Personinfo.class);
-        }
-    }
-
-    @Override
-    public HealthCheckResult checkHealth() {
-        return HealthCheckUtils.pingUrl(joinPaths(veilarbportefoljeUrl, "/internal/isAlive"), client);
-    }
-}
