+++ src/main/java/no/nav/veilarbperson/config/ClientConfig.java
@@ -22,8 +22,6 @@ import no.nav.veilarbperson.client.difi.DifiClientImpl;
 import no.nav.veilarbperson.client.difi.SbsServiceUser;
 import no.nav.veilarbperson.client.dkif.DkifClient;
 import no.nav.veilarbperson.client.dkif.DkifClientImpl;
-import no.nav.veilarbperson.client.egenansatt.EgenAnsattClient;
-import no.nav.veilarbperson.client.egenansatt.EgenAnsattClientImpl;
 import no.nav.veilarbperson.client.kodeverk.KodeverkClient;
 import no.nav.veilarbperson.client.kodeverk.KodeverkClientImpl;
 import no.nav.veilarbperson.client.nom.SkjermetClient;
@@ -36,8 +34,6 @@ import no.nav.veilarbperson.client.person.PersonClient;
 import no.nav.veilarbperson.client.person.PersonClientImpl;
 import no.nav.veilarbperson.client.veilarboppfolging.VeilarboppfolgingClient;
 import no.nav.veilarbperson.client.veilarboppfolging.VeilarboppfolgingClientImpl;
-import no.nav.veilarbperson.client.veilarbportefolje.VeilarbportefoljeClient;
-import no.nav.veilarbperson.client.veilarbportefolje.VeilarbportefoljeClientImpl;
 import no.nav.veilarbperson.client.veilarbregistrering.VeilarbregistreringClient;
 import no.nav.veilarbperson.client.veilarbregistrering.VeilarbregistreringClientImpl;
 import no.nav.veilarbperson.service.AuthService;
@@ -82,16 +78,6 @@ public class ClientConfig {
     }
 
     @Bean
-    @Deprecated
-    public VeilarbportefoljeClient veilarbportefoljeClient(AuthService authService) {
-        String url = isProduction()
-                ? createNaisAdeoIngressUrl(VEILARBPORTEFOLJE, true)
-                : createNaisPreprodIngressUrl(VEILARBPORTEFOLJE, "q1", true);
-
-        return new VeilarbportefoljeClientImpl(url, authService::getInnloggetBrukerToken);
-    }
-
-    @Bean
     public VeilarboppfolgingClient veilarboppfolgingClient(AuthService authService) {
         String url = isProduction()
                 ? createNaisAdeoIngressUrl(VEILARBOPPFOLGING, true)
@@ -114,12 +100,6 @@ public class ClientConfig {
                 : createDevAdeoIngressUrl(PAM_CV_API, true);
 
         return new PamClientImpl(url, systemUserTokenProvider::getSystemUserToken);
-    }
-
-    @Bean
-    @Deprecated
-    public EgenAnsattClient egenAnsattClient(EnvironmentProperties properties, StsConfig stsConfig) {
-        return new EgenAnsattClientImpl(properties.getEgenAnsattV1Endpoint(), stsConfig);
     }
 
     @Bean
