+++ src/main/java/no/nav/veilarbperson/controller/PersonController.java
@@ -4,13 +4,11 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.AllArgsConstructor;
 import no.nav.common.types.identer.Fnr;
 import no.nav.veilarbperson.client.difi.HarLoggetInnRespons;
-import no.nav.veilarbperson.client.person.domain.TpsPerson;
 import no.nav.veilarbperson.domain.*;
 import no.nav.veilarbperson.service.AuthService;
 import no.nav.veilarbperson.service.CvJobbprofilService;
-import no.nav.veilarbperson.service.PersonService;
+import no.nav.veilarbperson.service.PersonV2Service;
 import no.nav.veilarbperson.service.RegistreringService;
-import no.nav.veilarbperson.utils.PersonDataMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -20,8 +18,7 @@ import org.springframework.web.server.ResponseStatusException;
 @RequestMapping("/api/person")
 @AllArgsConstructor
 public class PersonController {
-
-    private final PersonService personService;
+    private final PersonV2Service personV2Service;
 
     private final AuthService authService;
 
@@ -33,10 +30,8 @@ public class PersonController {
     @Operation(summary = "Henter informasjon om en person",
             description = "Denne tjenesten gjør kall mot flere baktjenester: " +
                     "Kodeverk, organisasjonenhet_v2, Digitalkontaktinformasjon_v1, Person_v3, Egenansatt_v1")
-    public PersonData person(@PathVariable("fodselsnummer") Fnr fnr) {
-        authService.stoppHvisEksternBruker();
-        authService.sjekkLesetilgang(fnr);
-        return personService.hentFlettetPerson(fnr);
+    public PersonDataTPS person(@PathVariable("fodselsnummer") Fnr fnr) {
+        throw new ResponseStatusException(HttpStatus.GONE, "Bytt til v2 endepunkt");
     }
 
     @GetMapping("/aktorid")
@@ -49,22 +44,14 @@ public class PersonController {
     @GetMapping("/navn")
     @Operation(summary = "Henter navnet til en person")
     public PersonNavn navn(@RequestParam(value = "fnr", required = false) Fnr fnr) {
-        Fnr fodselsnummer = hentIdentForEksternEllerIntern(fnr);
-
-        authService.sjekkLesetilgang(fodselsnummer);
-
-        TpsPerson person = personService.hentPerson(fodselsnummer);
-        return PersonDataMapper.hentNavn(person);
+        throw new ResponseStatusException(HttpStatus.GONE, "Bytt til v2 endepunkt");
     }
 
     @GetMapping("/{fodselsnummer}/malform")
     @Operation(summary = "Henter målform til en person")
     public Malform malform(@PathVariable("fodselsnummer") Fnr fnr) {
-        authService.stoppHvisEksternBruker();
-        authService.sjekkLesetilgang(fnr);
-
-        TpsPerson person = personService.hentPerson(fnr);
-        return new Malform(person.getMalform());
+        throw new ResponseStatusException(HttpStatus.GONE,
+                "Bytt til v2 endepunkt");
     }
 
     @GetMapping("/{fodselsnummer}/tilgangTilBruker")
@@ -76,14 +63,12 @@ public class PersonController {
     public HarLoggetInnRespons harNivaa4(@PathVariable("fodselsnummer") Fnr fodselsnummer) {
         authService.stoppHvisEksternBruker();
         authService.sjekkLesetilgang(fodselsnummer);
-        return personService.hentHarNivaa4(fodselsnummer);
+        return personV2Service.hentHarNivaa4(fodselsnummer);
     }
 
     @GetMapping("/geografisktilknytning")
     public GeografiskTilknytning geografisktilknytning(@RequestParam(value = "fnr", required = false) Fnr fnr) {
-        Fnr fodselsnummer = hentIdentForEksternEllerIntern(fnr);
-        authService.sjekkLesetilgang(fodselsnummer);
-        return personService.hentGeografisktilknytning(fodselsnummer);
+        throw new ResponseStatusException(HttpStatus.GONE, "Bytt til v2 endepunkt");
     }
 
     @GetMapping("/cv_jobbprofil")
@@ -98,23 +83,4 @@ public class PersonController {
         return registreringService.hentRegistrering(fnr);
     }
 
-    // TODO: Det er hårete å måtte skille på ekstern og intern
-    //  Lag istedenfor en egen controller for interne operasjoner og en annen for eksterne
-    private Fnr hentIdentForEksternEllerIntern(Fnr queryParamFnr) {
-        Fnr fnr;
-
-        if (authService.erInternBruker()) {
-            if (queryParamFnr == null) {
-                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Mangler fnr");
-            }
-            fnr = queryParamFnr;
-        } else if (authService.erEksternBruker()) {
-            fnr = Fnr.of(authService.getInnloggerBrukerUid());
-        } else {
-            // Systembruker har ikke tilgang
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
-        }
-
-        return fnr;
-    }
 }
