+++ src/main/java/no/nav/veilarbperson/domain/PersonDataTPS.java
@@ -1,0 +1,10 @@
+package no.nav.veilarbperson.domain;
+
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+@Data
+@Accessors(chain = true)
+public class PersonDataTPS {
+    String kontonummer;
+}
