+++ src/test/java/no/nav/veilarbperson/service/PersonV2ServiceTest.java
@@ -3,8 +3,10 @@ package no.nav.veilarbperson.service;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import no.nav.common.client.norg2.Enhet;
 import no.nav.common.client.norg2.Norg2Client;
+import no.nav.common.featuretoggle.UnleashClient;
 import no.nav.common.sts.SystemUserTokenProvider;
 import no.nav.common.types.identer.Fnr;
+import no.nav.veilarbperson.client.difi.DifiCient;
 import no.nav.veilarbperson.client.dkif.DkifClient;
 import no.nav.veilarbperson.client.dkif.DkifKontaktinfo;
 import no.nav.veilarbperson.client.nom.SkjermetClient;
@@ -14,8 +16,8 @@ import no.nav.veilarbperson.client.pdl.PdlClient;
 import no.nav.veilarbperson.client.pdl.UserTokenProviderPdl;
 import no.nav.veilarbperson.client.pdl.domain.*;
 import no.nav.veilarbperson.client.person.PersonClient;
-import no.nav.veilarbperson.client.person.domain.RelasjonsBosted;
-import no.nav.veilarbperson.client.person.domain.TpsPerson;
+import no.nav.veilarbperson.client.pdl.domain.RelasjonsBosted;
+import no.nav.veilarbperson.client.person.TpsPerson;
 import no.nav.veilarbperson.config.PdlClientTestConfig;
 import no.nav.veilarbperson.domain.PersonNavnV2;
 import no.nav.veilarbperson.domain.PersonV2Data;
@@ -70,19 +72,21 @@ public class PersonV2ServiceTest extends PdlClientTestConfig {
         when(systemUserTokenProvider.getSystemUserToken()).thenReturn("SYSTEM_USER_TOKEN");
         when(norg2Client.hentTilhorendeEnhet(anyString(), any(), anyBoolean())).thenReturn(new Enhet());
         when(dkifClient.hentKontaktInfo(any())).thenReturn(new DkifKontaktinfo());
-        when(personClient.hentPerson(FNR)).thenReturn(new TpsPerson().setBarn(Collections.emptyList()));
+        when(personClient.hentPerson(FNR)).thenReturn(new TpsPerson().setKontonummer("123456789"));
         UserTokenProviderPdl tokenProvider = new UserTokenProviderPdl(() -> "test");
 
         personV2Service = new PersonV2Service(
                 pdlClient,
+                mock(DifiCient.class),
                 authService,
                 dkifClient,
                 norg2Client,
                 personClient,
+                mock(UnleashClient.class),
                 skjermetClient,
                 kodeverkService,
-                systemUserTokenProvider,
-                tokenProvider);
+                tokenProvider,
+                systemUserTokenProvider);
         person = hentPerson(FNR);
     }
